requirement.txt
flask==3.1.0
python-dotenv==1.0.0
psycopg2_binary==2.9.10
langchain==0.3.15
langchain-community==0.3.15
langchain-groq==0.2.3
langchain-huggingface==0.1.2
chromadb==0.6.3
pypdf==5.1.0
Werkzeug==3.1.3